content/ag_motioncontrol/generated_python_api/pyactiongraph/requirements.txt
pyserial>=3.5
pyserial-asyncio>=0.6
pyzmq>=25.0.1
tornado>=6.2; sys_platform == 'win32'
